refactor(showtimes): streamline date and time formatting
 Replace manual date and time string formatting with utility functions `formatDateYYYYMMDD` and `parseTime12Hour` to improve code maintainability and readability.
 This change simplifies the logic in the showtimes handler by utilizing the new utility functions from `timezone.js` to handle date and time conversions
 consistently across the application.
